Change imprime_capital_inicial parameter type to float

+++ src/prints.h
@@ -6,11 +6,10 @@ void imprime_pontilhado();
 
 void imprime_periodo(int ano_ini, int ano_fim);
 
-void imprime_capital_inicial(int capital_inicial);
+void imprime_capital_inicial(float capital_inicial);
 
 void imprime_resultado(const char *nome_arquivo, float resposta);
 
 void imprime_tam_janela(int janela);
-
 
 void imprime_nome_arquivo_janela(const char *nome_arquivo);
